Exception handling - delete

+++ src/main/java/com/springproject/course/services/CategoryService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.springproject.course.entities.Category;
 import com.springproject.course.repositories.CategoryRepository;
+import com.springproject.course.services.exceptions.DatabaseException;
 import com.springproject.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,14 @@ public class CategoryService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id))
+			throw new ResourceNotFoundException(id);
+		try {
+			repository.deleteById(id);			
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public Category update(Long id, Category obj) {
